requirements.txt
accelerate==1.3.0
aiohappyeyeballs
aiohttp==3.10.5
aiosignal==1.2.0
async-timeout
attrs==23.1.0
Bottleneck
Brotli 
certifi 
charset-normalizer 
contourpy
cycler
decord==0.6.0
dill 
einops==0.8.0
filelock 

fonttools==4.51.0
frozenlist==1.4.0
fsspec 
gmpy2 
huggingface_hub 
idna
importlib-metadata==7.0.1
importlib_resources
jieba 
Jinja2
joblib==1.4.2
kiwisolver 
Levenshtein==0.26.1
MarkupSafe 
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft 
mkl_random 
mpmath 
multidict==6.0.4
multiprocess==0.70.15
networkx 
numexpr 
numpy 
OpenCC==1.1.6
opencv-python
opencv-python-headless
packaging 
pandas
pillow
propcache==0.2.1
psutil 
py-cpuinfo
pyarrow 
pyarrow-hotfix 
pyparsing
PySocks 
python-dateutil 
pytz 
PyYAML 
RapidFuzz==3.11.0
regex==2024.11.6
requests 
rouge 
safetensors==0.5.2
scikit-learn==1.5.1
scipy
seaborn
sentencepiece==0.2.0
six 
sympy 
threadpoolctl==3.5.0
timm==0.9.12
tokenizers==0.20.3
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm 
transformers==4.45.2
triton==3.0.0
typing_extensions 
tzdata 
ultralytics
unicodedata2
urllib3
xxhash 
yarl==1.11.0
zipp==3.17.0
#/home/luoyx/InternVL/InternVL-main/flash_attn-2.6.1+cu118torch2.4cxx11abiFALSE-cp39-cp39-linux_x86_64.whl